+++ demos/browser/app/meeting/meeting.ts
@@ -167,8 +167,9 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver 
       new AsyncScheduler().start(
         async (): Promise<void> => {
           this.showProgress('progress-authenticate');
+          let meetingId: string = "";
           try {
-            await this.authenticate();
+            meetingId = await this.authenticate();
           } catch (error) {
             (document.getElementById(
               'failed-meeting'
@@ -180,7 +181,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver 
           }
           (document.getElementById(
             'meeting-id'
-          ) as HTMLSpanElement).innerHTML = `${this.meeting} (${this.region})`;
+          ) as HTMLSpanElement).innerHTML = `${meetingId} ${this.meeting} (${this.region})`;
           (document.getElementById('info-meeting') as HTMLSpanElement).innerHTML = this.meeting;
           (document.getElementById('info-name') as HTMLSpanElement).innerHTML = this.name;
           this.switchToFlow('flow-devices');
@@ -1050,7 +1051,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver 
     return value;
   }
 
-  async authenticate(): Promise<void> {
+  async authenticate(): Promise<string> {
     let joinInfo = (await this.joinMeeting()).JoinInfo;
     await this.initializeMeetingSession(
       new MeetingSessionConfiguration(joinInfo.Meeting, joinInfo.Attendee)
@@ -1058,6 +1059,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver 
     const url = new URL(window.location.href);
     url.searchParams.set('m', this.meeting);
     history.replaceState({}, `${this.meeting}`, url.toString());
+    return joinInfo.Meeting.meetingId;
   }
 
   log(str: string): void {